backend-fastapi/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
httpx
requests